vxl 1.15.0.20110428-r31940 (reduced)
This tree was extracted from upstream vxl by the following shell code.

url=https://vxl.svn.sourceforge.net/svnroot/vxl/trunk
v=1.15.0.20110428
r=31940
paths="
  CMakeLists.txt
  config/cmake
  core/CMakeLists.txt
  core/testlib
  core/vnl
  core/vxl_config.h.in
  core/vxl_copyright.h
  core/vxl_version.h
  v3p/CMakeLists.txt
  v3p/netlib
  vcl
"
date=$(svn log -q -c$r $url |
       sed -n "/^r/ {s/[^|]*|[^|]*|//;p;}")
svn export -r$r $url vxl-$v-r$r &&
mkdir vxl-$v-r$r-reduced &&
(cd vxl-$v-r$r && tar c $paths) |
(cd vxl-$v-r$r-reduced && tar x)
echo "r$r date: $date"

+++ v3p/netlib/blas/caxpy.c
@@ -48,8 +48,7 @@ extern "C" {
         return 0;
     }
 /*<       if (abs(real(ca)) + abs(aimag(ca)) .eq. 0.0 ) return >*/
-    if ((r__1 = ca->r, dabs(r__1)) + (r__2 = r_imag(ca), dabs(r__2)) == (
-            float)0.) {
+    if ((r__1 = ca->r, dabs(r__1)) + (r__2 = r_imag(ca), dabs(r__2)) == 0.f) {
         return 0;
     }
 /*<       if(incx.eq.1.and.incy.eq.1)go to 20 >*/
